pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["Kyrylo-Ktl <kyrylo.ktl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
environs = "^9.3.5"
django-extensions = "^3.1.5"
Faker = "^10.0.0"
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
flake8-print = "^4.0.0"
djangorestframework-simplejwt = "^5.0.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
drf-yasg = "^1.20.0"
pytest = "^6.2.5"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
